+++ src/Common/Contracts/AmbientWeatherMetricsContracts.cs
@@ -11,28 +11,28 @@ public record AmbientWeatherMetricsPostRequest : IAmbientWeatherMetrics
 	public string? StationType { get; init; }
 	public string? PassKey { get; init; }
 	public DateTime? DateUtc { get; init; }
-	public int? WindDir { get; init; }
+	public short? WindDir { get; init; }
 	public float? WindSpeedMph { get; init; }
 	public float? WindGustMph { get; init; }
-	public int? WindGustDir { get; init; }
+	public short? WindGustDir { get; init; }
 	public float? MaxDailyGust { get; init; }
 	public float? WindSpdMph_Avg2m { get; init; }
-	public int? WindDir_Avg2m { get; init; }
+	public short? WindDir_Avg2m { get; init; }
 	public float? WindSpdMph_Avg10m { get; init; }
-	public int? WindDir_Avg10m { get; init; }
-	public int? WindGustMph_Interval { get; init; }
-	public int? Humidity { get; init; }
-	public int? HumidityIn { get; init; }
-	public int? Humidity1 { get; init; }
-	public int? Humidity2 { get; init; }
-	public int? Humidity3 { get; init; }
-	public int? Humidity4 { get; init; }
-	public int? Humidity5 { get; init; }
-	public int? Humidity6 { get; init; }
-	public int? Humidity7 { get; init; }
-	public int? Humidity8 { get; init; }
-	public int? Humidity9 { get; init; }
-	public int? Humidity10 { get; init; }
+	public short? WindDir_Avg10m { get; init; }
+	public short? WindGustMph_Interval { get; init; }
+	public short? Humidity { get; init; }
+	public short? HumidityIn { get; init; }
+	public short? Humidity1 { get; init; }
+	public short? Humidity2 { get; init; }
+	public short? Humidity3 { get; init; }
+	public short? Humidity4 { get; init; }
+	public short? Humidity5 { get; init; }
+	public short? Humidity6 { get; init; }
+	public short? Humidity7 { get; init; }
+	public short? Humidity8 { get; init; }
+	public short? Humidity9 { get; init; }
+	public short? Humidity10 { get; init; }
 	public float? TempF { get; init; }
 	public float? TempInF { get; init; }
 	public float? Temp1F { get; init; }
@@ -55,19 +55,19 @@ public record AmbientWeatherMetricsPostRequest : IAmbientWeatherMetrics
 	public float? TotalRain { get; init; }
 	public float? BaromRelIn { get; init; }
 	public float? BaromAbsIn { get; init; }
-	public int? UV { get; init; }
+	public short? UV { get; init; }
 	public float? SolarRadiation { get; init; }
-	public int? CO2 { get; init; }
-	public int? PM25 { get; init; }
+	public short? CO2 { get; init; }
+	public short? PM25 { get; init; }
 	public float? PM25_24h { get; init; }
-	public int? PM25_In { get; init; }
+	public short? PM25_In { get; init; }
 	public float? PM25_In_24h { get; init; }
-	public int? PM10_In { get; init; }
+	public short? PM10_In { get; init; }
 	public float? PM10_In_24h { get; init; }
-	public int? CO2_In { get; init; }
+	public short? CO2_In { get; init; }
 	public float? CO2_In_24h { get; init; }
 	public float? PM_In_Temp { get; init; }
-	public int? PM_In_Humidity { get; init; }
+	public short? PM_In_Humidity { get; init; }
 	public BoolEnum? Relay1 { get; init; }
 	public BoolEnum? Relay2 { get; init; }
 	public BoolEnum? Relay3 { get; init; }
@@ -88,22 +88,22 @@ public record AmbientWeatherMetricsPostRequest : IAmbientWeatherMetrics
 	public float? SoilTemp8 { get; init; }
 	public float? SoilTemp9 { get; init; }
 	public float? SoilTemp10 { get; init; }
-	public int? SoilHum1 { get; init; }
-	public int? SoilHum2 { get; init; }
-	public int? SoilHum3 { get; init; }
-	public int? SoilHum4 { get; init; }
-	public int? SoilHum5 { get; init; }
-	public int? SoilHum6 { get; init; }
-	public int? SoilHum7 { get; init; }
-	public int? SoilHum8 { get; init; }
-	public int? SoilHum9 { get; init; }
-	public int? SoilHum10 { get; init; }
+	public short? SoilHum1 { get; init; }
+	public short? SoilHum2 { get; init; }
+	public short? SoilHum3 { get; init; }
+	public short? SoilHum4 { get; init; }
+	public short? SoilHum5 { get; init; }
+	public short? SoilHum6 { get; init; }
+	public short? SoilHum7 { get; init; }
+	public short? SoilHum8 { get; init; }
+	public short? SoilHum9 { get; init; }
+	public short? SoilHum10 { get; init; }
 	public LeakSensorState? Leak1 { get; init; }
 	public LeakSensorState? Leak2 { get; init; }
 	public LeakSensorState? Leak3 { get; init; }
 	public LeakSensorState? Leak4 { get; init; }
-	public long? Lightning_Time { get; init; }
-	public int? Lightning_Day { get; init; }
+	public ulong? Lightning_Time { get; init; }
+	public short? Lightning_Day { get; init; }
 	public BoolEnum? Lightning_Distance { get; init; }
 	public BoolEnum? LowBattery { get; init; }
 	public BoolEnum? NormalBattery { get; init; }
@@ -142,28 +142,28 @@ public record AmbientWeatherMetricsPostRequest : IAmbientWeatherMetrics
 	public BoolEnum? BattSM4 { get; init; }
 	public BoolEnum? BattRain { get; init; }
 	public BoolEnum? Batt_Co2 { get; init; }
-	public double? DewPoint { get; init; }
-	public double? DewPoint1 { get; init; }
-	public double? DewPoint2 { get; init; }
-	public double? DewPoint3 { get; init; }
-	public double? DewPoint4 { get; init; }
-	public double? DewPoint5 { get; init; }
-	public double? DewPoint6 { get; init; }
-	public double? DewPoint7 { get; init; }
-	public double? DewPoint8 { get; init; }
-	public double? DewPoint9 { get; init; }
-	public double? DewPoint10 { get; init; }
-	public double? DewPointIn { get; init; }
-	public double? FeelsLike { get; init; }
-	public double? FeelsLike1 { get; init; }
-	public double? FeelsLike2 { get; init; }
-	public double? FeelsLike3 { get; init; }
-	public double? FeelsLike4 { get; init; }
-	public double? FeelsLike5 { get; init; }
-	public double? FeelsLike6 { get; init; }
-	public double? FeelsLike7 { get; init; }
-	public double? FeelsLike8 { get; init; }
-	public double? FeelsLike9 { get; init; }
-	public double? FeelsLike10 { get; init; }
-	public double? FeelsLikeIn { get; init; }
+	public float? DewPoint { get; init; }
+	public float? DewPoint1 { get; init; }
+	public float? DewPoint2 { get; init; }
+	public float? DewPoint3 { get; init; }
+	public float? DewPoint4 { get; init; }
+	public float? DewPoint5 { get; init; }
+	public float? DewPoint6 { get; init; }
+	public float? DewPoint7 { get; init; }
+	public float? DewPoint8 { get; init; }
+	public float? DewPoint9 { get; init; }
+	public float? DewPoint10 { get; init; }
+	public float? DewPointIn { get; init; }
+	public float? FeelsLike { get; init; }
+	public float? FeelsLike1 { get; init; }
+	public float? FeelsLike2 { get; init; }
+	public float? FeelsLike3 { get; init; }
+	public float? FeelsLike4 { get; init; }
+	public float? FeelsLike5 { get; init; }
+	public float? FeelsLike6 { get; init; }
+	public float? FeelsLike7 { get; init; }
+	public float? FeelsLike8 { get; init; }
+	public float? FeelsLike9 { get; init; }
+	public float? FeelsLike10 { get; init; }
+	public float? FeelsLikeIn { get; init; }
 }
